Moved all items into a seperate file.
