+++ cobbler/collections/files.py
@@ -19,7 +19,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 """
 
 from collections import collection
-from cobbler import item_file as file
+from items import item_file as file
 from cobbler import utils
 from cobbler.cexceptions import CX
 from cobbler.utils import _
