+++ cobbler/collections/repos.py
@@ -21,7 +21,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 import os.path
 
 from collections import collection
-from cobbler import item_repo as repo
+from items import item_repo as repo
 from cobbler import utils
 from cobbler.cexceptions import CX
 from cobbler.utils import _
