+++ cobbler/items/item_distro.py
@@ -20,7 +20,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
 
 import os
 
-from cobbler import item
+from items import item
 from cobbler import utils
 from cobbler.cexceptions import CX
 from cobbler.utils import _
